We are now in world! We can run around freely now without any issue. This however means nothing else works, which is fine. I wanted to first achieve "sandbox" status with the emulator, and now that we have we can move onto the next stage. The next stage is going to be building up the funcitonality of AlefLogin. MySQL will be the DB engine of choice, simply for it's ease of use and Poco already has builtin support for it. After AlefLogin is finished, we will move back to AlefWorld and start emulating it's proper functionality.

+++ AlefCommon/AlefPacketPool.h
@@ -17,7 +17,7 @@ public:
 	~AlefPacketPool();
 
 	void addPacketToPool(AlefPacket* pkt);
-	int	getSize() { return packetPool->size(); }
+	size_t	getSize() { return packetPool->size(); }
 
 private:
 	UInt64 packetID;
